fix(datastore): refactor InitialSyncOperation tests
